Split of_stats to use multiples files.
